pyproject.toml
[project]
name = "mood2music"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.2",
    "numpy>=1.26.4",
    "pandas>=2.2.2",
    "scikit-learn>=1.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0", # useful for testing FastAPI endpoints
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
